Moved drag related code into its own namespace.
