--- AnnoMapEditor/UI/Controls/DragEndedEvent.cs
@@ -1,16 +1,0 @@
-﻿using AnnoMapEditor.Utilities;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace AnnoMapEditor.UI.Controls
-{
-    public delegate void DragEndedEventHandler(object? sender, DragEndedEventArgs e);
-
-
-    public class DragEndedEventArgs
-    {
-    }
-}
