+++ AnnoMapEditor/UI/Controls/Dragging/DragEndedEvent.cs
@@ -1,0 +1,9 @@
+﻿namespace AnnoMapEditor.UI.Controls.Dragging
+{
+    public delegate void DragEndedEventHandler(object? sender, DragEndedEventArgs e);
+
+
+    public class DragEndedEventArgs
+    {
+    }
+}
